2.07b

+++ qemu_mode/build_qemu_support.sh
@@ -6,7 +6,7 @@
 # Written by Andrew Griffiths <agriffiths@google.com> and
 #            Michal Zalewski <lcamtuf@google.com>
 #
-# Copyright 2015 Google Inc. All rights reserved.
+# Copyright 2015, 2016 Google Inc. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -124,6 +124,8 @@ patch -p0 <patches/syscall.diff || exit 1
 
 echo "[+] Patching done."
 
+ORIG_CPU_TARGET="$CPU_TARGET"
+
 test "$CPU_TARGET" = "" && CPU_TARGET="`uname -m`"
 test "$CPU_TARGET" = "i686" && CPU_TARGET="i386"
 
@@ -152,35 +154,43 @@ ls -l ../afl-qemu-trace || exit 1
 
 echo "[+] Successfully created '../afl-qemu-trace'."
 
-echo "[*] Testing the build..."
+if [ "$ORIG_CPU_TARGET" = "" ]; then
 
-cd ..
+  echo "[*] Testing the build..."
 
-make >/dev/null || exit 1
+  cd ..
 
-gcc test-instr.c -o test-instr || exit 1
+  make >/dev/null || exit 1
 
-unset AFL_INST_RATIO
+  gcc test-instr.c -o test-instr || exit 1
 
-echo 0 | ./afl-showmap -m none -Q -q -o .test-instr0 ./test-instr || exit 1
-echo 1 | ./afl-showmap -m none -Q -q -o .test-instr1 ./test-instr || exit 1
+  unset AFL_INST_RATIO
 
-rm -f test-instr
+  echo 0 | ./afl-showmap -m none -Q -q -o .test-instr0 ./test-instr || exit 1
+  echo 1 | ./afl-showmap -m none -Q -q -o .test-instr1 ./test-instr || exit 1
 
-cmp -s .test-instr0 .test-instr1
-DR="$?"
+  rm -f test-instr
 
-rm -f .test-instr0 .test-instr1
+  cmp -s .test-instr0 .test-instr1
+  DR="$?"
 
-if [ "$DR" = "0" ]; then
+  rm -f .test-instr0 .test-instr1
 
-  echo "[-] Error: afl-qemu-trace instrumentation doesn't seem to work!"
-  exit 1
+  if [ "$DR" = "0" ]; then
+
+    echo "[-] Error: afl-qemu-trace instrumentation doesn't seem to work!"
+    exit 1
+
+  fi
+
+  echo "[+] Instrumentation tests passed. "
+  echo "[+] All set, you can now use the -Q mode in afl-fuzz!"
+
+else
+
+  echo "[!] Note: can't test instrumentation when CPU_TARGET set."
+  echo "[+] All set, you can now (hopefully) use the -Q mode in afl-fuzz!"
 
 fi
-
-echo "[+] Instrumentation tests passed. "
-
-echo "[+] All set, you can now use the -Q mode in afl-fuzz!"
 
 exit 0
